[AI Chat]: Add Model Icons to Regenerate Answer Menu
