+++ components/ai_chat/resources/common/constants.ts
@@ -6,3 +6,20 @@
 export const MAX_IMAGES = 64
 
 export const IGNORE_EXTERNAL_LINK_WARNING_KEY = 'IGNORE_EXTERNAL_LINK_WARNING'
+
+const modelIcons = {
+  'chat-automatic': 'product-brave-leo',
+  'chat-deepseek-r1': 'deepseek-color',
+  'chat-claude-instant': 'anthropic-color',
+  'chat-claude-haiku': 'anthropic-color',
+  'chat-claude-sonnet': 'anthropic-color',
+  'chat-qwen': 'qwen-color',
+  'chat-basic': 'meta-color',
+  'chat-vision-basic': 'meta-color',
+}
+
+const fallbackModelIcon = 'product-brave-leo'
+
+export function getModelIcon(modelKey: string): string {
+  return modelIcons[modelKey as keyof typeof modelIcons] ?? fallbackModelIcon
+}
