+++ components/ai_chat/resources/page/components/feature_button_menu/index.tsx
@@ -14,21 +14,9 @@ import { getLocale } from '$web-common/locale'
 import * as Mojom from '../../../common/mojom'
 import { useAIChat } from '../../state/ai_chat_context'
 import { useConversation } from '../../state/conversation_context'
+import { getModelIcon } from '../../../common/constants'
 import styles from './style.module.scss'
 import useHasConversationStarted from '../../hooks/useHasConversationStarted'
-
-const modelIcons = {
-  'chat-automatic': 'product-brave-leo',
-  'chat-deepseek-r1': 'deepseek-color',
-  'chat-claude-instant': 'anthropic-color',
-  'chat-claude-haiku': 'anthropic-color',
-  'chat-claude-sonnet': 'anthropic-color',
-  'chat-qwen': 'qwen-color',
-  'chat-basic': 'meta-color',
-  'chat-vision-basic': 'meta-color',
-}
-
-const fallbackModelIcon = 'product-brave-leo'
 
 export interface Props {
   setIsConversationsListOpen?: (value: boolean) => unknown
@@ -82,7 +70,7 @@ export default function FeatureMenu(props: Props) {
           >
             <div className={styles.menuItemWithIcon}>
               <div className={styles.modelIcon} data-key={model.key}>
-                <Icon name={modelIcons[model.key as keyof typeof modelIcons] ?? fallbackModelIcon} />
+                <Icon name={getModelIcon(model.key)} />
               </div>
               <div className={styles.menuText}>
                 <div>{model.displayName}</div>
